refactor(hyperswarm): extract negentropy
