+++ services/mediators/hyperswarm/src/mediator-state.ts
@@ -1,15 +1,5 @@
 import type { HyperswarmConnection } from 'hyperswarm';
-import type { Operation } from '@mdip/gatekeeper/types';
-import type {
-    OperationSyncStore,
-    SyncStoreCursor,
-} from './db/types.js';
-import type {
-    NegentropyWindowEngine,
-    NegentropyWindowSnapshot,
-    NegentropyWindowStats,
-    ReconciliationWindow,
-} from './negentropy/adapter.js';
+import type { OperationSyncStore } from './db/types.js';
 import type {
     NegotiatedPeerCapabilities,
     SyncMode,
@@ -57,34 +47,6 @@ export interface MalformedPeerState {
     lastRejectLogAt: number;
 }
 
-export interface PeerSyncSession {
-    sessionId: string;
-    peerKey: string;
-    initiator: boolean;
-    windows: ReconciliationWindow[];
-    windowIndex: number;
-    windowId: string | null;
-    currentWindowStats: NegentropyWindowStats | null;
-    currentWindowSnapshot: NegentropyWindowSnapshot | null;
-    currentWindowEngine: NegentropyWindowEngine | null;
-    startedAt: number;
-    lastActivity: number;
-    pendingHaveIds: Set<string>;
-    pendingNeedIds: Set<string>;
-    unresolvedNeedIds: Set<string>;
-    unresolvedOperations: Map<string, Operation>;
-    rounds: number;
-    maxRounds: number;
-    reconciliationComplete: boolean;
-    localClosed: boolean;
-    receivedPushIds: Set<string>;
-    receivedKnownPushIds: Set<string>;
-    provenStoredPushIds: Set<string>;
-    receivedPushMaxCursor: SyncStoreCursor | null;
-    remoteWindowCappedByRecords: boolean;
-    remoteWindowLastCursor: SyncStoreCursor | null;
-}
-
 export interface ConnectionInfoOptions {
     connection: HyperswarmConnection;
     peerKey: string;
@@ -126,45 +88,5 @@ export function createConnectionInfo(options: ConnectionInfoOptions): Connection
         legacyTransportQuarantined: false,
         inboundBuffer: Buffer.alloc(0),
         inboundReceiveChain: Promise.resolve(),
-    };
-}
-
-export interface PeerSyncSessionOptions {
-    sessionId: string;
-    peerKey: string;
-    initiator: boolean;
-    maxRounds: number;
-    now?: number;
-}
-
-export function createPeerSyncSessionState(options: PeerSyncSessionOptions): PeerSyncSession {
-    const now = options.now ?? Date.now();
-
-    return {
-        sessionId: options.sessionId,
-        peerKey: options.peerKey,
-        initiator: options.initiator,
-        windows: [],
-        windowIndex: 0,
-        windowId: null,
-        currentWindowStats: null,
-        currentWindowSnapshot: null,
-        currentWindowEngine: null,
-        startedAt: now,
-        lastActivity: now,
-        pendingHaveIds: new Set<string>(),
-        pendingNeedIds: new Set<string>(),
-        unresolvedNeedIds: new Set<string>(),
-        unresolvedOperations: new Map<string, Operation>(),
-        rounds: 0,
-        maxRounds: options.maxRounds,
-        reconciliationComplete: false,
-        localClosed: false,
-        receivedPushIds: new Set<string>(),
-        receivedKnownPushIds: new Set<string>(),
-        provenStoredPushIds: new Set<string>(),
-        receivedPushMaxCursor: null,
-        remoteWindowCappedByRecords: false,
-        remoteWindowLastCursor: null,
     };
 }
